responding to PR and bringing in comments in code from commentSamples branch
