+++ PluginExampleProject/src/main/scala/com/alpine/plugin/samples/ver1_0/DBNumericFeatureTransformer.scala
@@ -32,7 +32,7 @@ class DBNumericFeatureTransformerSignature extends OperatorSignature[
   DBNumericFeatureTransformerRuntime] {
   def getMetadata(): OperatorMetadata = {
     new OperatorMetadata(
-      name = "Sample - DB Numeric Feature Transformer",
+      name = "Sample - DB Transform",
       category = "Plugin Sample - DB",
       author = "Sung Chung",
       version = 1,
