+++ PluginExampleProject/src/test/scala/com/alpine/plugin/samples/ver1_0/BadDataReportingPluginJobTest.scala
@@ -3,7 +3,10 @@ package com.alpine.plugin.samples.ver1_0
 import com.alpine.plugin.core.utils.HdfsParameterUtils
 import com.alpine.plugin.test.mock.OperatorParametersMock
 import com.alpine.plugin.test.utils.{ParameterMockUtil, SimpleAbstractSparkJobSuite}
+import org.junit.runner.RunWith
+import org.scalatest.junit.JUnitRunner
 
+@RunWith(classOf[JUnitRunner])
 class BadDataReportingPluginJobTest extends SimpleAbstractSparkJobSuite {
   import com.alpine.plugin.core.spark.utils.TestSparkContexts._
   test("Test that null data is removed") {
